+++ CHANGELOG.md
@@ -12,11 +12,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - Add-PnPTeamsTeam: new cmdlet that creates a Teams team for the current, Office 365 group connected, site collection
 - Added Get-PnPTenantId to retrieve the current tenant id.
 - ConvertTo-PnPClientSidePage: Added support for enforcing the specified target page folder via `-TargetPageFolderOverridesDefaultFolder`
+- ConvertTo-PnPClientSidePage: Added support for Delve blog page modernization via the `-DelveBlogPage` and `-DelveKeepSubTitle ` parameters
 
 ### Changed
 
 - Fixes issue with App Only with certificate and context cloning. Now Apply-PnPTenantTemplate works as expected.
 - Added CorrelationId and TimeStampUtc to output of Get-PnPException which can help in analyzing ULS entries.
+- ConvertTo-PnPClientSidePage: The `-Identity` parameter now also accepts the item id as value to find a page
 
 ### Contributors